+++ apps/hydrogen-app/hydrogen.config.ts
@@ -1,0 +1,28 @@
+import {CookieSessionStorage, defineConfig} from '@shopify/hydrogen/config';
+
+export default defineConfig({
+  shopify: {
+    defaultCountryCode: 'US',
+    defaultLanguageCode: 'EN',
+    privateStorefrontToken:
+      // @ts-ignore
+      Oxygen?.env?.PRIVATE_STOREFRONT_API_TOKEN,
+    storeDomain:
+      // @ts-ignore
+      Oxygen?.env?.PUBLIC_STORE_DOMAIN || 'oxygenator.myshopify.com',
+    storefrontToken:
+      // @ts-ignore
+      Oxygen?.env?.PUBLIC_STOREFRONT_API_TOKEN ||
+      '70faab4b482211c4167f94181a4ba4ed',
+    storefrontApiVersion: '2022-07',
+    // @ts-ignore
+    storefrontId: Oxygen?.env?.PUBLIC_STOREFRONT_ID,
+  },
+  session: CookieSessionStorage('__session', {
+    path: '/',
+    httpOnly: true,
+    secure: import.meta.env.PROD,
+    sameSite: 'Strict',
+    maxAge: 60 * 60 * 24 * 30,
+  }),
+});