+++ apps/hydrogen-app/vite.config.js
@@ -1,0 +1,8 @@
+import {defineConfig} from 'vite';
+import hydrogen from '@shopify/hydrogen/plugin';
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  plugins: [hydrogen()],
+  optimizeDeps: {include: ['@headlessui/react']},
+});