+++ apps/nextjs-app/src/features/demo/pages/DemoPage.tsx
@@ -1,5 +1,5 @@
-import { sayHello } from '@your-org/core-lib';
-import { AsyncMessage, Message } from '@your-org/ui-lib';
+import { sayHello } from '@we-make-websites/core-lib';
+import { AsyncMessage, Message } from '@we-make-websites/ui-lib';
 import { useTranslation } from 'next-i18next';
 import { NextSeo } from 'next-seo';
 import Image from 'next/image';
@@ -23,9 +23,13 @@ export const DemoPage: FC = () => {
         <h3>I'm the web-app</h3>
         <Jumbotron />
         <ul>
-          <li>{`Foo says: ${sayHello('World')} from @your-org/core-lib`}</li>
+          <li>{`Foo says: ${sayHello(
+            'World'
+          )} from @we-make-websites/core-lib`}</li>
           <li>
-            <Message message={'Bar react component from @your-org/ui-lib'} />
+            <Message
+              message={'Bar react component from @we-make-websites/ui-lib'}
+            />
           </li>
           <li>
             <AsyncMessage apiUrl={'/api/hello'} />