+++ packages/ts-utils/CHANGELOG.md
@@ -1,4 +1,4 @@
-# @your-org/ts-utils
+# @we-make-websites/ts-utils
 
 ## 1.1.1
 
@@ -13,6 +13,6 @@
 
 ### Minor Changes
 
-- [#2713](https://github.com/belgattitude/nextjs-monorepo-example/pull/2713) [`e279d984`](https://github.com/belgattitude/nextjs-monorepo-example/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Move typescript utils to @your-org/ts-utils
+- [#2713](https://github.com/belgattitude/nextjs-monorepo-example/pull/2713) [`e279d984`](https://github.com/belgattitude/nextjs-monorepo-example/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Move typescript utils to @we-make-websites/ts-utils
 
 - [#2713](https://github.com/belgattitude/nextjs-monorepo-example/pull/2713) [`e279d984`](https://github.com/belgattitude/nextjs-monorepo-example/commit/e279d984d01937264b077a111863b38d15fcb1d0) Thanks [@belgattitude](https://github.com/belgattitude)! - Example ts-utils packages as esm with jest and vitest